move cardBox from component to container, paving way to promote it'
