+++ client/components/cardBox.js
@@ -1,29 +1,29 @@
-'use strict'
-
-import React from 'react';
-import ReactDOM from 'react-dom';
-import CardTitle from './cardTitle';
-import PackageDestinations from '../containers/packageDestinations';
-import { Provider } from 'react-redux';
-import {createStore, applyMiddleware } from 'redux';
-import reducers from '../reducers';
-import promise from 'redux-promise';
-
-const createStoreWithMiddleware = applyMiddleware(
-	promise
-	)(createStore);
-
-class CardBox extends React.Component {
-	render() {
-		console.log('line 18 components/cardBox.js params of package/:package_name', this.props.params.package_name);
-		return(			
-			<div className='seven-view'>
-				<CardTitle title={this.props.params.package_name} className='card-title' /> 
-				<PackageDestinations  package_name={this.props.params.package_name} className='row' />
-			</div>
-		)
-	}
-}
+// 'use strict'
+// import React from 'react';
+// import ReactDOM from 'react-dom'
+// import CardTitle from './cardTitle'
+// import PackageDestinations from '../containers/packageDestinations'
+// import { Provider } from 'react-redux';
+// import {createStore, applyMiddleware } from 'redux';
+// import reducers from '../reducers';
+// import promise from 'redux-promise';
 
 
-export default CardBox;
+// const createStoreWithMiddleware = applyMiddleware(
+// 	promise
+// 	)(createStore);
+
+// class CardBox extends React.Component {
+// 	render() {
+// 		console.log('line 18 components/cardBox.js params of package/:package_name', this.props.params.package_name);
+// 		return(			
+// 			<div className='seven-view'>
+// 				<CardTitle title={this.props.params.package_name} className='card-title' /> 
+// 				<PackageDestinations  package_name={this.props.params.package_name} className='row' />
+// 			</div>
+// 		)
+// 	}
+// }
+
+
+// export default CardBox;
